email: Move `UserConfirmEmail` struct to its corresponding module
