+++ src/controllers/user/me.rs
@@ -6,7 +6,6 @@ use crate::controllers::frontend_prelude::*;
 use crate::controllers::helpers::*;
 
 use crate::controllers::helpers::pagination::{Paginated, PaginationOptions};
-use crate::email::UserConfirmEmail;
 use crate::models::{
     CrateOwner, Email, Follow, NewEmail, OwnerKind, User, Version, VersionOwnerAction,
 };
@@ -296,4 +295,29 @@ pub async fn update_email_notifications(app: AppState, req: BytesRequest) -> App
         ok_true()
     })
     .await
+}
+
+pub struct UserConfirmEmail<'a> {
+    pub user_name: &'a str,
+    pub domain: &'a str,
+    pub token: &'a str,
+}
+
+impl crate::email::Email for UserConfirmEmail<'_> {
+    const SUBJECT: &'static str = "Please confirm your email address";
+
+    fn body(&self) -> String {
+        // Create a URL with token string as path to send to user
+        // If user clicks on path, look email/user up in database,
+        // make sure tokens match
+
+        format!(
+            "Hello {user_name}! Welcome to crates.io. Please click the
+link below to verify your email address. Thank you!\n
+https://{domain}/confirm/{token}",
+            user_name = self.user_name,
+            domain = self.domain,
+            token = self.token,
+        )
+    }
 }
